requirements-phase1.txt
# Phase 1 Requirements (Python 3.14 compatible)
# Core infrastructure only - database, API clients, config, logging, CLI

# Core
python-dotenv==1.1.1
pyyaml==6.0.3

# Database
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
alembic==1.15.2

# API Clients (Sync/Async/WebSocket)
requests==2.32.5
httpx==0.27.2
websockets==15.0.1
aiohttp==3.13.0

# Utilities (Datetime + Validation/CLI/Auth/Scheduling)
python-dateutil==2.9.0.post0
pytz==2025.1
pydantic==2.12.0
click==8.1.7
cryptography==46.0.2

# Development Tools (Testing/Linting/Static Analysis)
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-cov==7.0.0
black==24.8.0
flake8==7.3.0
mypy==1.12.0

# Monitoring and Logging
structlog==25.4.0

# Phase 2+ packages (not needed yet)
# pandas==2.3.3  # Will install in Phase 3-4 when needed
# numpy==2.3.3   # Will install in Phase 3-4 when needed
# apscheduler==3.11.0  # Will install in Phase 2
# spacy==3.8.7  # Phase 8 - not Python 3.14 compatible yet
